#70-record-tuya-data

+++ src/main/java/com/example/SmartBuildingBackend/service/implementation/AqaraServiceImplemetation.java
@@ -331,7 +331,7 @@ public class AqaraServiceImplemetation implements AqaraService {
                     DEFAULT_TEMPERATURE = Integer.parseInt(temperature);
                     // input LogValue to store value
                     logValueDto.setTimeStamp(node.get("timeStamp").asLong());
-                    logValueDto.setValueResponse(node.get("value").asLong());
+                    logValueDto.setValueResponse(node.get("value").asDouble());
                     Long valueId = valueService.getValueByName("temperature");
                     logValueService.addLogValue(equipmentDto.getEquipmentId(), valueId, logValueDto);
 
@@ -340,7 +340,7 @@ public class AqaraServiceImplemetation implements AqaraService {
                     result.put("humidity", valueNode.asText().substring(0, 2));
                     // input LogValue to store value
                     logValueDto.setTimeStamp(node.get("timeStamp").asLong());
-                    logValueDto.setValueResponse(node.get("value").asLong());
+                    logValueDto.setValueResponse(node.get("value").asDouble());
                     Long valueId = valueService.getValueByName("humidity");
                     logValueService.addLogValue(equipmentDto.getEquipmentId(), valueId, logValueDto);
                 }
